requirements.txt
amqp==5.2.0
annotated-types==0.7.0
ansicon==1.89.0
arrow==1.3.0
asgiref==3.8.1
billiard==4.2.1
blessed==1.20.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.1.2
django-celery-results==2.5.1
django-ninja==1.3.0
django-picklefield==3.2
django-prometheus==2.3.1
django-q==1.3.9
django-role-permissions==3.2.0
djangorestframework==3.15.2
gevent==24.10.3
greenlet==3.1.1
idna==3.10
jinxed==1.3.0
kombu==5.4.2
prometheus_client==0.21.0
prompt_toolkit==3.0.48
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
pytz==2024.2
redis==3.5.3
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
zope.event==5.0
zope.interface==7.1.1
django-redis==5.4.0
hiredis
